Update unpacking script
Multiple HEAD references

Multiple siva files with multiple references

Multiple references with unreachable paths

+++ src/file_utils/unpacking.py
@@ -23,7 +23,7 @@ logger.addHandler(console)
 logger.addHandler(errors)
 
 siva_path = "./siva/latest"
-repos_path = "./repositories"
+repos_path = "./repos"
 index_file_path = "./index.csv"
 
 logger.debug("Making repo dir")
@@ -56,6 +56,11 @@ with open(index_file_path) as f:
         langs = row[3].lower().split(',')
         if not any(l in langs for l in ["python", "java"]):
             logger.debug("{}/{} do not have python/java source code".format(author, name))
+            logger.info("Skipping line {} {}/{}".format(line_no, author, name))
+            continue
+
+        if os.path.exists("{repo_path}/{author}/{name}".format(repo_path=repos_path, author=author, name=name)):
+            logger.debug("{}/{} already exists".format(author, name))
             logger.info("Skipping line {} {}/{}".format(line_no, author, name))
             continue
         # ------------------
@@ -97,14 +102,73 @@ with open(index_file_path) as f:
         logger.debug("Calculating number of siva files with HEAD references")
         head_sivas = os.listdir(repo_dir)
         if len(head_sivas) > 1:
-            logger.critical("Problem in line {}".format(line_no))
-            logger.critical(
+            # TODO: refactor. Could be a general case. Particular is when there is only one siva and one ref.
+            logger.warning(
                 "More than one siva file with HEAD for repository {}/{} ({})".format(author, name, len(head_sivas)))
-            logger.critical("Siva files with HEADs {}".format(head_sivas))
-            logger.info("Skipping")
-            logger.debug("Removing {}".format(repo_dir))
-            shutil.rmtree(repo_dir, ignore_errors=True)
-            continue
+            logger.debug("Siva files with HEADs {}".format(head_sivas))
+
+            wd = os.getcwd()
+            repo_ref_map = {}
+            reference = []
+            logger.debug("Searching siva file and HEAD ref")
+            try:
+                for _siva in head_sivas:
+                    refs = os.listdir(repo_dir + "/{}/.git/refs/heads/HEAD".format(_siva))
+                    for _ref in refs:
+                        if _ref in repo_ref_map:
+                            logger.critical("Problem in line {}".format(line_no))
+                            logger.critical("Reference in present in more than one siva file")
+                            logger.critical("2 of them: {}".format([repo_ref_map[_ref], _siva]))
+                            raise StopIteration()
+
+                        repo_ref_map[_ref] = _siva
+
+                    logger.debug("Changing dir to {}".format(repo_dir + "/{}/".format(_siva)))
+                    os.chdir(repo_dir + "/{}/".format(_siva))
+                    remotes = subprocess.check_output(["git", "remote", "-v"]).decode('utf-8')
+                    ref_line = re.search(r'.*{}'.format(url), remotes, re.MULTILINE)
+                    if ref_line:
+                        reference.append(ref_line.group(0).split('\t')[0])
+                        logger.info("Found reference {}".format(reference[-1]))
+
+                    logger.debug("Changing dir to {}".format(wd))
+                    os.chdir(wd)
+
+            except StopIteration:
+                logger.info("Skipping")
+                continue
+
+            if not reference:
+                logger.critical("Problem in line {}".format(line_no))
+                logger.critical("No HEAD reference found for repository {}/{}".format(author, name))
+                logger.critical("No reference found for url {}".format(url))
+
+                # Should be already in wd
+                logger.debug("Removing {}".format(repo_dir))
+                shutil.rmtree(repo_dir, ignore_errors=True)
+                continue
+
+            ref = None
+            for r in reference:
+                try:
+                    logger.debug("Found pair siva/ref {}/{}".format(repo_ref_map[r], r))
+                except KeyError:
+                    logger.warning("Ref {} was not found".format(r))
+                    continue
+                else:
+                    siva = repo_ref_map[r]
+                    ref = r
+                    break
+
+            if not ref:
+                logger.critical("Problem in line {}".format(line_no))
+                logger.critical("Reference was not found for repository {}/{}".format(author, name))
+                logger.debug("Removing {}".format(repo_dir))
+                shutil.rmtree(repo_dir, ignore_errors=True)
+                continue
+            logger.debug("Changing dir to {}".format(repo_dir + "/{}/".format(siva)))
+            os.chdir(repo_dir + "/{}/".format(siva))
+
         elif len(head_sivas) == 0:
             logger.critical("Problem in line {}".format(line_no))
             logger.critical("No siva file with HEAD for repository {}/{}".format(author, name))
@@ -128,7 +192,11 @@ with open(index_file_path) as f:
                 # open config/remotes and search for <url>, then corresponding <ref>
                 # git remote -v to get pair <reference> <url>
                 remotes = subprocess.check_output(["git", "remote", "-v"]).decode('utf-8')
-                ref = re.search(r'.*{}'.format(url), remotes, re.MULTILINE).group(0).split('\t')[0]
+                ref_line = re.search(r'.*{}'.format(url), remotes, re.MULTILINE)
+                if ref_line:
+                    ref = ref_line.group(0).split('\t')[0]
+                else:
+                    ref = None
                 if not ref:
                     logger.critical("Problem in line {}".format(line_no))
                     logger.critical(
@@ -149,32 +217,32 @@ with open(index_file_path) as f:
 
             logger.info("Found reference {}".format(ref))
 
-            # checkout <ref>
-            logger.info("Checking out reference")
-            subprocess.run(["git", "checkout", "refs/heads/HEAD/{}".format(ref)], stdout=subprocess.DEVNULL,
-                           stderr=subprocess.DEVNULL)
+        # checkout <ref>
+        logger.info("Checking out reference")
+        subprocess.run(["git", "checkout", "refs/heads/HEAD/{}".format(ref)], stdout=subprocess.DEVNULL,
+                       stderr=subprocess.DEVNULL)
 
-            logger.debug("Changing dir to {}".format(wd))
-            os.chdir(wd)
-            # rm .git
-            logger.debug("Cleaning .git directory")
-            shutil.rmtree(repo_dir + "/{}/.git".format(siva), ignore_errors=True)
-            # mv author/repo/<siva>/* to author/repo/
-            logger.info("Moving files from siva directory to repository")
-            logger.debug("Moving dir {} to {}".format(repo_dir + "/{}".format(siva), repo_dir))
-            src = repo_dir + "/{}/".format(siva)
-            for content in os.listdir(src):
-                try:
-                    if os.path.islink(src + content):
-                        shutil.copy2(src + content, repo_dir, follow_symlinks=False)
-                    else:
-                        shutil.move(src + content, repo_dir)
-                except OSError as e:
-                    logger.critical(
-                        "Error handling {} in repository {}/{} siva file {} with ref {}".format(content, author, name,
-                                                                                                siva, ref))
-                    logger.critical("{}".format(e))
-                    logger.debug("Ignoring...")
-            # rm siva directory
-            logger.debug("Cleaning siva directory")
-            shutil.rmtree(repo_dir + "/{}/".format(siva), ignore_errors=True)
+        logger.debug("Changing dir to {}".format(wd))
+        os.chdir(wd)
+        # rm .git
+        logger.debug("Cleaning .git directory")
+        shutil.rmtree(repo_dir + "/{}/.git".format(siva), ignore_errors=True)
+        # mv author/repo/<siva>/* to author/repo/
+        logger.info("Moving files from siva directory to repository")
+        logger.debug("Moving dir {} to {}".format(repo_dir + "/{}".format(siva), repo_dir))
+        src = repo_dir + "/{}/".format(siva)
+        for content in os.listdir(src):
+            try:
+                if os.path.islink(src + content):
+                    shutil.copy2(src + content, repo_dir, follow_symlinks=False)
+                else:
+                    shutil.move(src + content, repo_dir)
+            except OSError as e:
+                logger.critical(
+                    "Error handling {} in repository {}/{} siva file {} with ref {}".format(content, author, name,
+                                                                                            siva, ref))
+                logger.critical("{}".format(e))
+                logger.debug("Ignoring...")
+        # rm siva directory
+        logger.debug("Cleaning siva directory")
+        shutil.rmtree(repo_dir + "/{}/".format(siva), ignore_errors=True)
